Migrating current progress to repo.
Current progress includes having a table that accepts user input. Each row is an entry, consisting of a name, grade value, and weight value. The calculate button will find the current grade based on the weights provided and based on their overall grade based on the missing weight (if any).